Fix int16 spell ID limit with items (click, procs, etc)
See issue #7 (Did we resolve it yet?)

+++ common/item_struct.h
@@ -51,7 +51,7 @@
 **
 */
 struct ItemEffect_Struct {
-	int16	Effect;
+	int32	Effect;
 	uint8	Type;
 	uint8	Level;
 	uint8	Level2;
